requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
dataclasses==0.7
decorator==4.4.0
defusedxml==0.6.0
einops==0.1.0
entrypoints==0.3
func-timeout==4.3.5
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
html5lib==0.9999999
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
monty==3.0.2
more-itertools==7.2.0
mpi4py==3.0.3
mpmath==1.1.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
palettable==3.3.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyDispatcher==2.0.5
pydot==1.4.1
Pygments==2.4.2
pylint==2.4.2
pymatgen==2019.10.16
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
spglib==1.14.1.post0
sympy==1.4
tabulate==0.8.5
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.2.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
